Dictionary drawer will constantly read value if the member type was a KVPList
Since now DictionaryDrawer works for all IDictionary<K, V>, it means it will also be applied to KVPList<K, V> (which implemented IDictionary<K, V>) - KVPList is used in our serialization data. Since in DictionaryDrawer we're internally using KVPList to represent Dictionary and since we don't read/write per frame (causes tons of allocation), this causes the issue where if you're viewing the serialization data and modifying the corresponding values in the behaviour from the inspector, the values in the dictionary doesn't change! DictionaryDrawer can't know it should re-read in this case. The solution was just to handle KVPList as a special case, if we have a KVPList member, we always read because it's free, doesn't cost any allocations.