Gradient Fields
For the most part this should work. It does throw some nasty exceptions
when not using the try catch block, but after applying the try catch, it
seems to be working fine.

Also I don't know why but the size of the field seems way larger than it
should be.