Refactored type-drawer detection, introduced TypeDrawerMapper.
This was one of the very dark corners of the framework. Instead of that hacky, ugly, nightmare to maintainable, error prone, buggy detection code, now we have a very clear, straight-forward mapping system that is very explicit and deterministic. Only drawback is that now you have to explicitly register/map your drawer. See the corresponding example file.