Improved AnimVar Functionality
AnimVar now works with ints and AnimatorControllerParameter objects in
addition to the strings it used to work with. With ints, it will
automatically save the hash value or ID of the parameter. The
AnimatorControllerParameter object can also be used.

The type of variable can also now be filtered via the attribute. For
example:

[AnimVar(ParameterType.Bool)]
public int BoolVar;

This would only allow BoolVar to be a boolean variable in the controller.

The filter is a mask, so a group of parameter types can be used. Like so:

[AnimVar(ParameterType.Float | ParameterType.Int)]
public string FloatOrIntVar;

Other general functionality of AnimVar has not been altered.

To do this, however, I had to create an extra ParameterType enum (as
Unity's AnimatorControllerParameterType is not flag/mask friendly) and an
alternate AttributeDrawer class that only had one generic type.