VFWSettings and BetterEditorPrefs are now movable
VFWSettings and BetterEditorPrefs can now work no matter where they are
stored. As long as it is under the project directory, it should be
locatable.

Also instead of showing a error message when it is not possible to
create an instance under */Editor/Vexe/ScriptableAssets it now
automatically creates in the Assets root folder, which the developer can
move wherever they please.